add onboarding/ dir and extract wizard-setup logic
